Node pools reducer
