+++ src/actions/actionTypes.js
@@ -123,6 +123,10 @@ export const CLUSTERS_LOAD_NODEPOOLS_SUCCESS =
   'CLUSTERS_LOAD_NODEPOOLS_SUCCESS';
 export const CLUSTERS_LOAD_NODEPOOLS_ERROR = 'CLUSTERS_LOAD_NODEPOOLS_ERROR';
 
+// Node Pools
+export const NODEPOOLS_LOAD_SUCCESS = 'NODEPOOLS_LOAD_SUCCESS';
+export const NODEPOOLS_LOAD_ERROR = 'NODEPOOLS_LOAD_ERROR';
+
 export const NODEPOOL_PATCH = 'NODEPOOL_PATCH';
 export const NODEPOOL_PATCH_ERROR = 'NODEPOOL_PATCH_ERROR';
