+++ src/reducers/clusterReducer.js
@@ -60,25 +60,6 @@ export default function clusterReducer(
         items,
       };
 
-    case types.CLUSTERS_LOAD_NODEPOOLS_SUCCESS: {
-      return {
-        ...state,
-        items: {
-          ...state.items,
-          [action.clusterId]: {
-            ...state.items[action.clusterId],
-            nodePools: action.nodePools,
-          },
-        },
-      };
-    }
-
-    case types.CLUSTERS_LOAD_NODEPOOLS_ERROR:
-      return {
-        ...state,
-        errorLoading: true,
-      };
-
     case types.CLUSTER_LOAD_DETAILS_SUCCESS:
       items[action.cluster.id] = Object.assign(
         {},
@@ -110,6 +91,19 @@ export default function clusterReducer(
       return {
         ...state,
         items,
+      };
+
+    case types.CLUSTERS_LOAD_NODEPOOLS_SUCCESS:
+      return {
+        ...state,
+        lastUpdated: Date.now(),
+        items: {
+          ...state.items,
+          [action.clusterId]: {
+            ...state.items[action.clusterId],
+            nodePools: [...action.nodePools],
+          },
+        },
       };
 
     case types.CLUSTER_LOAD_STATUS_SUCCESS:
@@ -254,74 +248,6 @@ export default function clusterReducer(
           },
         },
       };
-
-    case types.NODEPOOL_PATCH: {
-      // Node pools in state.
-      const nodePools = [...state.items[action.clusterId].nodePools];
-
-      // Find the node pool we want to modify.
-      const oldNodePool = nodePools.find(np => np.id === action.nodePool.id);
-
-      // Find its index
-      const indexOfOldNodePool = nodePools.findIndex(
-        np => np.id === action.nodePool.id
-      );
-
-      // Merge node pool with new data coming from the action (payload).
-      const newNodePool = { ...oldNodePool, ...action.payload };
-
-      // Remove old new pool and add new one.
-      const newNodePools = (nodePools, index) => {
-        return [
-          ...nodePools.slice(0, index),
-          ...nodePools.slice(index + 1),
-          newNodePool,
-        ];
-      };
-
-      return {
-        ...state,
-        lastUpdated: Date.now(),
-        items: {
-          ...state.items,
-          [action.clusterId]: {
-            ...state.items[action.clusterId],
-            nodePools: newNodePools(nodePools, indexOfOldNodePool),
-          },
-        },
-      };
-    }
-
-    case types.NODEPOOL_PATCH_ERROR: {
-      // Node pools in state.
-      const nodePools = [...state.items[action.clusterId].nodePools];
-
-      // Find the index of the node pool we want to modify
-      const indexOfOldNodePool = nodePools.findIndex(
-        np => np.id === action.nodePool.id
-      );
-
-      // Remove old new pool and add new one.
-      const newNodePools = (nodePools, index) => {
-        return [
-          ...nodePools.slice(0, index),
-          ...nodePools.slice(index + 1),
-          action.nodePool,
-        ];
-      };
-
-      return {
-        ...state,
-        lastUpdated: Date.now(),
-        items: {
-          ...state.items,
-          [action.clusterId]: {
-            ...state.items[action.clusterId],
-            nodePools: newNodePools(nodePools, indexOfOldNodePool),
-          },
-        },
-      };
-    }
 
     default:
       return state;
